propagate S/N parameters to FeaFi and PeaPi
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@3834 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/FeatureFinderAlgorithmWatershed.h
@@ -126,7 +126,7 @@ namespace OpenMS
         //---------------------------------------------------------------------------
         debug_ = param_.getValue("debug").toBool();
         mz_sampling_ = (DoubleReal)(param_.getValue("mz_sampling"));
-        peaks_ = (map_->getMaxMZ() - map_->getMinMZ()) / mz_sampling_;
+        peaks_ = UInt((map_->getMaxMZ() - map_->getMinMZ()) / mz_sampling_);
         
         //---------------------------------------------------------------------------
       	//RESAMPLE AND BUILD MAIN DATASTRUCTURE
